watermarking_manager/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
requests
timm